Separate Text and TextComponent
